+++ src/main/java/org/farm/fireflyserver/domain/senior/web/mapper/SeniorMapper.java
@@ -3,7 +3,6 @@ package org.farm.fireflyserver.domain.senior.web.mapper;
 import org.farm.fireflyserver.domain.senior.persistence.entity.Senior;
 import org.farm.fireflyserver.domain.senior.web.dto.request.RegisterSeniorDto;
 import org.mapstruct.Mapper;
-import org.mapstruct.factory.Mappers;
 
 @Mapper(componentModel = "spring")
 public interface SeniorMapper {
